+++ komarovTelegrambot/src/main/java/by/komarov/komarovTelegrambot/Processor/HelpProcessor.java
@@ -1,0 +1,12 @@
+package by.komarov.komarovTelegrambot.Processor;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class HelpProcessor implements Processor {
+
+    @Override
+    public String run() {
+        return "Sorry, currently all operators are busy";
+    }
+}
