+++ komarovTelegrambot/src/main/java/by/komarov/komarovTelegrambot/Processor/SettingsProcessor.java
@@ -1,0 +1,13 @@
+package by.komarov.komarovTelegrambot.Processor;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class SettingsProcessor implements Processor {
+
+    @Override
+    public String run() {
+
+        return "Nothing you can set now, sorry...";
+    }
+}
